pyproject.toml
[tool.ruff]
line-length = 120
select = ["E", "F", "UP", "N", "C", "B", "I001", "S", "Q"]
ignore = [
       "S101", # allow use of assert
       # FIXME: S105, S106, S608 and S113 should be addressed
       "S105", "S106", # possible hardcoded password
       "S608", # possible SQL injection
       "S113", # requests call without timeout
       "N801", # class names
       "N802", # function names
       "N818", # exception names
       "B017", # pytest.raises(Exception)
       "C901", # complexity
]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.ruff.format]
# Prefer single quotes over double quotes
quote-style = "single"

[tool.poetry]
name = "toucan-connectors"
version = "5.0.0"
description = "Toucan Toco Connectors"
authors = ["Toucan Toco <dev@toucantoco.com>"]
license = "BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
Authlib = "^1.0.1"
cached-property = "^1.5.2"
Jinja2 = "^3.0.3"
jq = "^1.2.2"
pydantic = "^2.4.2"
requests = "^2.28.0"
tenacity = "^8.0.1"
toucan-data-sdk = "^7.6.0"

# Dependencies for extras
adobe-analytics = {version = "^1.2.3", optional = true}
bearer = {version = "3.1.0", optional = true}
oauthlib = {version = "3.2.2", optional = true}
requests-oauthlib = {version = "1.3.1", optional = true}
awswrangler = {version = "^3.0.0", optional = true}
pyodbc = {version = "^4", optional = true}
clickhouse-driver = {version = ">=0.2.3,<1.0", optional = true}
dataiku-api-client = {version = ">=9.0.1,<13.0.0", optional = true}
elasticsearch = {version = ">=7.11.0,<8", optional = true}
facebook-sdk = {version = "^3.1.0", optional = true}
python-graphql-client = {version = ">=0.4.3,<1.0", optional = true}
google-api-python-client = {version = "^2", optional = true}
oauth2client = {version = "^4.1.3", optional = true}
googleads = {version = ">=34,<35", optional = true}
google-cloud-bigquery = {version = ">=3,<4", extras = ["bqstorage", "pandas"], optional = true}
PyMySQL = {version = "^1.0.2", optional = true}
gspread = {version = "^5.4.0", optional = true}
xmltodict = {version = ">=0.13.0,<1.0", optional = true}
pymongo = {version = ">=3.12.0", optional = true}
tctc-odata = {version = ">=0.3,<1.0", optional = true}
cx-Oracle = {version = "^8.3.0", optional = true}
openpyxl = {version = "^3.0.9", optional = true}
psycopg2 = {version = "^2.7.4", optional = true}
redshift-connector = {version = "^2.0.907", optional = true}
lxml = {version = "^4.6.5", optional = true}
PyJWT = {version = ">=1.5.3,<3", optional = true}
simplejson = {version = "^3.17.6", optional = true}
pyhdb = {version = ">=0.3.4,<1.0", optional = true}
zeep = {version = "^4.1.0", optional = true}
snowflake-connector-python = {version = ">=2.7.12,<4.0.0", optional = true}
pyarrow = {version = "*", optional = true}
toucan-client = {version = "^1.0.1", optional = true}
peakina = {version = ">=0.11,<0.13", optional = true}
hubspot-api-client = {version = ">=7.4,<9.0", optional = true}
aiohttp = "^3.8.4"

[tool.poetry.dev-dependencies]
Authlib = "^1.0.1"
aioresponses = ">=0.7.3,<1.0"
click = "^8.1.7"
cryptography = ">=40.0.2"
docker = "^6.1.3"
mock = "^5.1.0"
pytest-aiohttp = "^1.0.5"
pytest-asyncio = ">=0.19.0,<1"
pytest-cov = "^4.1.0"
pytest = "^7.4.3"
pytest-mock = "^3.12.0"
pytest-rerunfailures = "^12.0"
python-slugify = "<7"
PyYAML = ">=6.0.1,<7"
responses = ">=0.21.0,<1"
psycopg2 = "^2.9.9"
xmltodict = ">=0.13.0,<1"
python-graphql-client = ">=0.4.3,<1"
clickhouse-driver = ">=0.2.3,<1"
lxml = "4.9.3"
zeep = "^4.2.1"
mypy = "^1.7"
pandas-stubs = "^2.1.1.230928"
types-requests = "^2.31.0.6"
types-simplejson = "^3.19.0.2"
types-python-slugify = "^8.0.0.3"
types-pyopenssl = "^23.3.0.0"

[tool.poetry.extras]
adobe = ["adobe-analytics"]
awsathena = ["awswrangler"]
azure_mssql = ["pyodbc"]
clickhouse = ["clickhouse-driver"]
dataiku = ["dataiku-api-client"]
elasticsearch = ["elasticsearch"]
facebook = ["facebook-sdk"]
github = ["python-graphql-client"]
google_analytics = ["google-api-python-client", "oauth2client"]
google_adwords = ["google-adwords"]
google_big_query = ["google-cloud-bigquery"]
google_cloud_mysql = ["PyMySQL"]
google_my_business = ["google-api-python-client"]
google_sheets = ["google-api-python-client"]
google_spreadsheet = ["gspread", "oauth2client"]
http_api = ["oauthlib", "requests-oauthlib", "xmltodict"]
hubspot = ["hubspot-api-client"]
mongo = ["pymongo"]
mssql = ["pyodbc"]
mssql_TLSv1_0 = ["pyodbc"]
mysql = ["PyMySQL"]
odata = ["oauthlib", "requests-oauthlib", "tctc_odata"]
odbc = ["pydobc"]
oracle_sql = ["cx_Oracle"]
net_explorer = ["openpyxl"]
postgres = ["psycopg2"]
Redshift = ["redshift_connector", "lxml"]
sap_hana = ["pyhdb"]
soap = ["zeep", "lxml"]
snowflake = ["snowflake-connector-python", "PyJWT", "pyarrow"]
toucan_toco = ["toucan_client"]

# All
all = [
    "adobe-analytics",
    "awswrangler",
    "bearer",
    "clickhouse-driver",
    "cx-Oracle",
    "dataiku-api-client",
    "elasticsearch",
    "facebook-sdk",
    "google-api-python-client",
    "google-cloud-bigquery",
    "googleads",
    "gspread",
    "hubspot-api-client",
    "lxml",
    "oauth2client",
    "oauthlib",
    "openpyxl",
    "peakina",
    "psycopg2",
    "pyarrow",
    "pyhdb",
    "PyJWT",
    "pymongo",
    "PyMySQL",
    "pyodbc",
    "python-graphql-client",
    "redshift-connector",
    "requests-oauthlib",
    "simplejson",
    "snowflake-connector-python",
    "tctc-odata",
    "toucan-client",
    "xmltodict",
    "zeep"
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.8"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
plugins = "pydantic.mypy"
follow_imports = "silent"
show_error_codes = true
warn_no_return = true
warn_unused_configs = true
warn_unused_ignores = true
files = [
      "toucan_connectors/toucan_connector.py",
      "toucan_connectors/hubspot_private_app/hubspot_connector.py",
      "toucan_connectors/snowflake/snowflake_connector.py",
      "toucan_connectors/peakina/peakina_connector.py"
]
